Вынесение конструкторов древовидного вида в общую логику
