+++ drevo/templates/drevo/constructors/tree_constructor.html
@@ -10,25 +10,32 @@
 {% endblock %}
 
 {% block jsfiles %}
-    <script src="{% static 'drevo/js/constructors/algorithm_constructor.js' %}"></script>
+    <script src="{% static 'drevo/js/constructors/tree_constructor.js' %}"></script>
     <script src="{% static 'drevo/js/get_cookie.js' %}"></script>
+    {% if type_of_zn == 'algorithm' %}
+        <script src="{% static 'drevo/js/constructors/algorithm_constructor.js' %}"></script>
+    {% else %}
+        <script src="{% static 'drevo/js/constructors/document_constructor.js' %}"></script>
+    {% endif %}
 {% endblock %}
 
 {% block content %}
 
-    <a href="{% url 'constructor_tree' %}?type_of_zn=algorithm" class="btn btn-outline-info mb-3">Назад</a>
+    <a href="{% url 'znaniya_for_constructor' %}?type_of_zn=algorithm" class="btn btn-outline-info mb-3">Назад</a>
     <h1>{{ title }}</h1>
     <div>
       {% if relative_znaniya %}
-        {% build_constructor_tree queryset=relative_znaniya show_complex=True is_constructor_type=type %}
+        {% build_constructor_tree queryset=relative_znaniya show_complex=True is_constructor_type='algorithm' %}
       {% else %}
-          {% build_constructor_tree queryset=znanie show_complex=True %}
+          {% build_constructor_tree queryset=znanie show_complex=True is_constructor_type='algorithm' %}
       {% endif %}
     </div>
     <div id="div-param" class="button-block">
         <div class="mt-3">
-            <input class="btn btn-outline-success mx-1" id="open_algorithm" type="button" value="Открыть">
-            <input class="btn btn-outline-primary" id="btn_create_copy" type="button" value="Создать копию">
+            <input class="btn btn-outline-success mx-1" id="open_zn" type="button" value="Открыть">
+            {% if type_of_zn == 'algorithm' %}
+                <input class="btn btn-outline-primary" id="btn_create_copy" type="button" value="Создать копию">
+            {% endif %}
         </div>
     </div>
 
@@ -72,6 +79,10 @@
                                     <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001zm-.644.766a.5.5 0 0 0-.707 0L1.95 11.756l-.764 3.057 3.057-.764L14.44 3.854a.5.5 0 0 0 0-.708l-1.585-1.585z"/>
                                 </svg>
                             </i>
+                        </div>
+                        <label for="order">Номер связи:</label>
+                        <div class="d-flex align-items-center gap-3">
+                            <input class="form-control" id="order">
                         </div>
                         <input id="save_new_relation" class="btn btn-success" style="margin-top: 7px" value="Сохранить">
                     </form>
@@ -156,42 +167,45 @@
         </div>
     </div>
 
-    <!-- Модальное окно для отображения ошибок алгоритма -->
-    <div class="modal fade" id="algorithm_errors_modal">
-        <div class="modal-dialog modal-dialog-centered modal-lg">
-            <div class="modal-content bg-white text-dark">
-                <div class="modal-header" style="justify-content: center;">
-                    <h3 class="modal-title text-center">Ошибки при проверке алгоритма!</h3>
-                    <button type="button" class="btn-close" style="position: absolute; right: 1rem;" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body" style="margin-left: 15px">
-                    <div id="less_than_min_block" hidden>
-                        <h5>Знания, в которых недостаточно внутренних связей:</h5>
-                        <div id="less_than_min"></div>
+    {% if type_of_zn == 'algorithm' %}
+        <!-- Модальное окно для отображения ошибок алгоритма -->
+        <div class="modal fade" id="algorithm_errors_modal">
+            <div class="modal-dialog modal-dialog-centered modal-lg">
+                <div class="modal-content bg-white text-dark">
+                    <div class="modal-header" style="justify-content: center;">
+                        <h3 class="modal-title text-center">Ошибки при проверке алгоритма!</h3>
+                        <button type="button" class="btn-close" style="position: absolute; right: 1rem;" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div class="modal-body" style="margin-left: 15px">
+                        <div id="less_than_min_block" hidden>
+                            <h5>Знания, в которых недостаточно внутренних связей:</h5>
+                            <div id="less_than_min"></div>
+                        </div>
                     </div>
                 </div>
             </div>
         </div>
-    </div>
 
-    <!-- Модальное окно для создания копии алгоритма с заданным именем -->
-    <div class="modal fade" id="algorithm_create_copy_modal" tabindex="-1" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="deleteElementModalLabel">Создать копию</h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body">
-                    <input type="text" id="name_for_copy" class="form-control mb-2" placeholder="Введите название алгоритма" aria-label="Введите название" aria-describedby="basic-addon2">
-                    <span id="algorithm-error" class="text-danger fs-6">• Копия данного алгоритма уже существует</span>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-success js-okay-successful" onclick="copy({{ main_zn_id }})">Создать</button>
+        <!-- Модальное окно для создания копии алгоритма с заданным именем -->
+        <div class="modal fade" id="algorithm_create_copy_modal" tabindex="-1" aria-hidden="true">
+            <div class="modal-dialog modal-dialog-centered">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title" id="deleteElementModalLabel">Создать копию</h5>
+                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div class="modal-body">
+                        <input type="text" id="name_for_copy" class="form-control mb-2" placeholder="Введите название алгоритма" aria-label="Введите название" aria-describedby="basic-addon2">
+                        <span id="algorithm-error" class="text-danger fs-6">• Копия данного алгоритма уже существует</span>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-success js-okay-successful" onclick="copy({{ main_zn_id }})">Создать</button>
+                    </div>
                 </div>
             </div>
         </div>
-    </div>
+    {% endif %}
+
     {#  Корректное отображение select в модальном окне  #}
     <style>
         .select2-container{
